apps/shopify/requirements.txt
requests>=2.28.0
python-dotenv>=0.20.0
pydantic>=1.9.0 